Extract DefaultFilterFn to new file.
